BUGFIX: File extension names could conflict with methods
This patch simply changes the file extensions, preserving the "." (period) at
the beginning of the extension.

The Hirb gem, used to output the table in the gem's output, essentially calls
object.send(field_name) to output each row's (Author's) column. In the case of
a file extension of "test", this was sending author.test. Test is a private
method on the Ruby Object class, thus Hirb was inadvertently calling "test"
instead of outputting what we expect.

Now, the column names are ".test", ".js", etc., which is still readable.